+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - ODPresentation Writer : Support for fill in RichText - @Progi1984 GH-79
 - ODPresentation Writer : Support for border style in RichText - @Progi1984 GH-79
 - ODPresentation Writer : Support for Area Chart - @Progi1984 GH-82
+- PowerPoint2007 Writer : Support for Area Chart - @Progi1984 GH-82
 
 ### Bugfix
 - PSR-0 via composer broken - @Progi1984 GH-51